+++ bakedanuki/bakedanuki-util/python/bd_util/maya/ui/binding/__init__.py
@@ -1,5 +1,6 @@
 # coding: utf-8
 from .enum_binding import MayaEnumBinding
+from .enum_definition import read_enum_definition
 from .enum_plug import MayaEnumPlugStore, MayaEnumPlugView
 from .enum_plug_binding import MayaEnumPlugBinding
 from .enum_plug_resolver import MayaEnumPlug, resolve_enum_plug
@@ -33,6 +34,7 @@ __all__ = [
     "MayaEnumPlugStore",
     "MayaEnumPlugView",
     "resolve_enum_plug",
+    "read_enum_definition",
     "MayaBoolPlugsBinding",
     "MayaFloatPlugsBinding",
     "MayaPlugTargetState",